+++ extensions/homeassistant/src/components/battery/utils.ts
@@ -1,4 +1,6 @@
+import { PrimaryIconColor } from "@components/state/utils";
 import { State } from "@lib/haapi";
+import { Color, Image } from "@raycast/api";
 
 export function sortBatteries(states: State[] | undefined) {
   if (!states) {
@@ -25,6 +27,48 @@ export function sortBatteries(states: State[] | undefined) {
   return sortedStates;
 }
 
+const batterLevelIcons: string[] = [
+  "battery-00.png",
+  "battery-10.png",
+  "battery-20.png",
+  "battery-30.png",
+  "battery-40.png",
+  "battery-50.png",
+  "battery-60.png",
+  "battery-70.png",
+  "battery-80.png",
+  "battery-90.png",
+  "battery-100.png",
+];
+
+export function getBatteryIconSourceForValue(value: number | undefined): string {
+  const fallback = "battery-100.png";
+  if (value === undefined) {
+    return fallback;
+  }
+  if (!isNaN(value)) {
+    const level = Math.floor(value / 10);
+    const levelIcon = batterLevelIcons[level];
+    if (levelIcon) {
+      const src = levelIcon;
+      return src;
+    }
+  }
+  return fallback;
+}
+
+export function getBatteryIconFromState(state: State): Image.ImageLike | undefined {
+  const v = parseFloat(state.state);
+  const src = getBatteryIconSourceForValue(v);
+  let tintColor = PrimaryIconColor;
+  if (v <= 20) {
+    tintColor = Color.Red;
+  } else if (v <= 30) {
+    tintColor = Color.Yellow;
+  }
+  return { source: src, tintColor: tintColor };
+}
+
 export function filterBatteries(
   states: State[] | undefined,
   options?: { onFilterState?: (value: number) => void },
@@ -40,6 +84,11 @@ export function filterBatteries(
     );
   }
   return batteries;
+}
+
+export function getBatteryStateValue(state: State): number {
+  const val = parseFloat(state.state);
+  return Number.isNaN(val) ? 0 : val;
 }
 
 export function batteryStateValue(state: State) {