+++ extensions/homeassistant/CHANGELOG.md
@@ -6,6 +6,7 @@
 - Add Entities Menu Bar command to allow to add specific entities in the Menu Bar
 - Add Lights Menu Bar command
 - Add Covers Menu Bar command
+- Add Batteries Menu Bar command
 - Notifications menu now contains battery low battery states as well as updates
 
 ## [Modernize] - 2023-07-16